Move records types to the new `record` module in aggregator
