+++ mithril-aggregator/src/database/provider/mod.rs
@@ -22,17 +22,3 @@ pub use single_signature::*;
 pub use stake_pool::*;
 #[cfg(test)]
 pub use test_helper::*;
-
-pub(crate) fn read_signed_entity_beacon_column<U: sqlite::RowIndex + Clone>(
-    row: &sqlite::Row,
-    column_index: U,
-) -> String {
-    // TODO: We need to check first that the cell can be read as a string first
-    // (e.g. when beacon json is '{"network": "dev", "epoch": 1, "immutable_file_number": 2}').
-    // If it fails, we fallback on reading the cell as an integer (e.g. when beacon json is '5').
-    // Maybe there is a better way of doing this.
-    match row.try_read::<&str, _>(column_index.clone()) {
-        Ok(value) => value.to_string(),
-        Err(_) => (row.read::<i64, _>(column_index)).to_string(),
-    }
-}
