+++ mithril-aggregator/src/database/provider/stake_pool.rs
@@ -1,80 +1,20 @@
-use anyhow::Context;
-use async_trait::async_trait;
-use chrono::{DateTime, Utc};
-use sqlite::Value;
 use std::iter::repeat;
 use std::ops::Not;
 use std::sync::Arc;
 
-use mithril_common::{
-    entities::{Epoch, PartyId, Stake, StakeDistribution},
-    StdResult,
-};
+use anyhow::Context;
+use async_trait::async_trait;
+use chrono::Utc;
+use sqlite::Value;
+
+use mithril_common::entities::{Epoch, PartyId, Stake, StakeDistribution};
+use mithril_common::StdResult;
 use mithril_persistence::sqlite::{
-    EntityCursor, HydrationError, Projection, Provider, SourceAlias, SqLiteEntity,
-    SqliteConnection, WhereCondition,
+    EntityCursor, Provider, SourceAlias, SqLiteEntity, SqliteConnection, WhereCondition,
 };
 use mithril_persistence::store::{adapter::AdapterError, StakeStorer};
 
-/// Stake pool as read from Chain.
-#[derive(Debug, PartialEq)]
-pub struct StakePool {
-    /// Pool Id
-    stake_pool_id: PartyId,
-
-    /// Total stake of this pool.
-    stake: u64,
-
-    /// Epoch at which this pool is valid.
-    epoch: Epoch,
-
-    /// DateTime of the record creation.
-    created_at: DateTime<Utc>,
-}
-
-impl SqLiteEntity for StakePool {
-    fn hydrate(row: sqlite::Row) -> Result<Self, HydrationError>
-    where
-        Self: Sized,
-    {
-        let epoch_int = row.read::<i64, _>(2);
-        let datetime = &row.read::<&str, _>(3);
-        let stake = row.read::<i64, _>(1);
-
-        let stake_pool = Self {
-            stake_pool_id: row.read::<&str, _>(0).to_string(),
-            stake: u64::try_from(stake).map_err(|e| {
-                HydrationError::InvalidData(format!(
-                    "Could not cast the StakePool.stake from internal db I64 → U64. Error: '{e}'.",
-                ))
-            })?,
-            epoch: Epoch(epoch_int.try_into().map_err(|e| {
-                HydrationError::InvalidData(format!(
-                    "Could not cast i64 ({epoch_int}) to u64. Error: '{e}'"
-                ))
-            })?),
-            created_at: DateTime::parse_from_rfc3339(datetime)
-                .map_err(|e| {
-                    HydrationError::InvalidData(format!(
-                        "Could not turn string '{datetime}' to rfc3339 Datetime. Error: {e}"
-                    ))
-                })?
-                .with_timezone(&Utc),
-        };
-
-        Ok(stake_pool)
-    }
-
-    fn get_projection() -> Projection {
-        let mut projection = Projection::default();
-        projection.add_field("stake_pool_id", "{:stake_pool:}.stake_pool_id", "text");
-        projection.add_field("stake", "{:stake_pool:}.stake", "integer");
-        projection.add_field("epoch", "{:stake_pool:}.epoch", "integer");
-        projection.add_field("created_at", "{:stake_pool:}.created_at", "text");
-
-        projection
-    }
-}
+use crate::database::record::StakePool;
 
 /// Simple [StakePool] provider.
 pub struct StakePoolProvider<'client> {
@@ -294,8 +234,9 @@ impl StakeStorer for StakePoolStore {
 mod tests {
     use sqlite::Connection;
 
-    use super::*;
     use crate::database::provider::apply_all_migrations_to_db;
+
+    use super::*;
 
     pub fn setup_stake_db(
         connection: &SqliteConnection,
